pyproject.toml
[tool.poetry]
name = "python-django"
version = "0.1.0"
description = ""
authors = ["Aleksei Savrasov <alekseisavrasov0@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
djangorestframework = "^3.15.1"
django-filter = "^24.2"
django-debug-toolbar = "^4.3.0"
pillow = "^10.3.0"
gunicorn = "^22.0.0"
sentry-sdk = "^2.3.1"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
